# Remove precompile statements