These were once necessary, but now they're doing more harm than good.

The `precompile_model` functions are safe to replace with a warning because they have no user-visible effect. 

## Code

```Julia
using JuMP, HiGHS
model = Model(HiGHS.Optimizer)
set_silent(model)
@variable(model, x >= 0)
@variable(model, 0 <= y <= 3)
@objective(model, Min, 12x + 20y)
@constraint(model, c1, 6x + 8y >= 100)
@constraint(model, c2, 7x + 12y >= 120)
optimize!(model)
```

## Before

```julia
(base) oscar@MacBookPro precompile % time julia --project=. main.jl
julia --project=. main.jl  2.73s user 0.45s system 101% cpu 3.119 total
(base) oscar@MacBookPro precompile % time julia --project=. main.jl
julia --project=. main.jl  2.76s user 0.41s system 101% cpu 3.111 total
(base) oscar@MacBookPro precompile % time julia --project=. main.jl
julia --project=. main.jl  2.73s user 0.41s system 102% cpu 3.077 total
```

## This PR

```julia
(base) oscar@MacBookPro precompile % time julia --project=. main.jl            
julia --project=. main.jl  2.24s user 0.42s system 101% cpu 2.623 total
(base) oscar@MacBookPro precompile % time julia --project=. main.jl
julia --project=. main.jl  2.24s user 0.37s system 102% cpu 2.560 total
(base) oscar@MacBookPro precompile % time julia --project=. main.jl
julia --project=. main.jl  2.28s user 0.39s system 102% cpu 2.620 total
```

There are also now better tools like PrecompileTools.jl so that we can actually build and solve problems in each package if needed.