@blegat 2025-02-20
JuMP and HiGHS both have precompile statements so that's a bit misleading. Maybe also include benchmarks with pure-MOI and a solver that does not have precompile statemtents