@odow 2025-02-20
Here's SCS:

Old
```
julia --project=old scs.jl  13.78s user 0.57s system 100% cpu 14.327 total
julia --project=old scs.jl  13.80s user 0.54s system 100% cpu 14.336 total
```
New
```
julia --project=. scs.jl  13.80s user 0.53s system 99% cpu 14.329 total
julia --project=. scs.jl  13.80s user 0.51s system 100% cpu 14.285 total
```

There's no `PrecompileTools`, and these existing recompilation statements did not help because we've switched the type that JuMP uses.

I guess the action item in addition to this PR is to add precompilation things to each solver package.